Crate eq_volatility

Equilibrium Volatility Pallet

Equilibrium's Volatility Pallet is a Substrate module that calculates and stores prices volatility and prices correlation for all system currencies. This data is used for fees and other system parameters calculation

Structs

Corellation
Pallet storage - price correlations for every pair of currencies

GenesisConfig
Genesis config for the module, allow to build genesis storage.

LastUpdate
Pallet storage - timestamps of updates for every Currency in each PricePeriod

Module

Prices
Pallet storage - vectors of prices for every Currency for each PricePeriod

Volatility
Pallet storage - volatilities for every Currency

Enums

Call
Dispatchable calls.

PricePeriod
A time period for which price is aggregated

Traits

Trait
Substrate pallet configuration trait

VolatilityGetter
Volatility, correlation and covariance storage adapter